tests: Fix unreliable package installation
The generic package module does not work reliably when the package manager's
cache is outdated, because (unlike most specific modules) it does not support
the update_cache option.

We probably shouldn't trigger an update_cache as part of the role itself, since
this would be redundant when multiple roles install packages. Thus, we assume
the playbook using this role does this, or the cache is kept up to date by
other means.